requirements.txt
pyrogram
python-dotenv
flask
gunicorn
